requirements.txt
moviepy==1.0.3
tqdm==4.66.2
openai-whisper==20231117
translate==3.6.1
gTTS==2.5.3
pydub==0.25.1